requirements.txt
supabase==2.9.1
python-dotenv==1.0.1
groq==0.11.0
google-generativeai==0.8.3
httpx==0.27.0
flask==3.0.3
feedparser==6.0.11
